BD/sim/requirements.txt
psycopg2-binary==2.9.3
prometheus_client==0.16.0